Remove support for Rails < 6
