+++ lib/data_migrate/status_service.rb
@@ -16,50 +16,37 @@ module DataMigrate
     end
 
     def dump(stream)
-      unless @connection.table_exists?(table_name)
-        stream.puts "Data migrations table does not exist yet."
-        return
-      end
-      sql = "SELECT version FROM #{DataMigrate::DataMigrator.schema_migrations_table_name}"
-      db_list = ActiveRecord::Base.connection.select_values(sql)
-      output(stream, db_list)
+      output(stream)
     end
 
     private
 
     def table_name
-      DataMigrate::DataMigrator.schema_migrations_table_name
+      DataMigrate::DataSchemaMigration.table_name
     end
 
-    def output(stream, db_list)
-      stream.puts "#{"Status".center(8)}  #{"Migration ID".ljust(14)}  Migration Name"
+    def output(stream)
+      unless DataMigrate::DataSchemaMigration.table_exists?
+        stream.puts "Data migrations table does not exist yet."
+        return
+      end
+
+      # output
+      stream.puts "\ndatabase: #{database}\n\n"
+      stream.puts "#{'Status'.center(8)}  #{'Migration ID'.ljust(14)}  Migration Name"
       stream.puts "-" * 50
-      list =  migration_files(db_list) + migration_list(db_list)
-      list.sort! {|line1, line2| line1[1] <=> line2[1]}
-      list.each do |file|
-        stream.puts "#{file[0].center(8)}  #{file[1].ljust(14)}  #{file[2]}"
+      db_list.each do |status, version, name|
+        stream.puts "#{status.center(8)}  #{version.ljust(14)}  #{name}"
       end
       stream.puts
     end
 
-    def migration_list(db_list)
-      list = []
-      db_list.each do |version|
-        list << ["up", version, "********** NO FILE *************"]
-      end
-      list
+    def database
+      ActiveRecord::Base.connection_db_config.database
     end
 
-    def migration_files(db_list)
-      file_list = []
-      Dir.foreach(File.join(root_folder, DataMigrate.config.data_migrations_path)) do |file|
-        # only files matching "20091231235959_some_name.rb" pattern
-        if match_data = DataMigrate::DataMigrator.match(file)
-          status = db_list.delete(match_data[1]) ? "up" : "down"
-          file_list << [status, match_data[1], match_data[2].humanize]
-        end
-      end
-      file_list
+    def db_list
+      DataMigrate::DataMigrator.migrations_status
     end
   end
 end
