+++ lib/data_migrate/status_service_six.rb
@@ -1,0 +1,11 @@
+module DataMigrate
+  module StatusServiceSix
+    private
+
+    def database
+      ActiveRecord::Base.connection_config[:database]
+    end
+  end
+end
+
+DataMigrate::StatusService.prepend DataMigrate::StatusServiceSix
