+++ lib/data_migrate/tasks/data_migrate_tasks.rb
@@ -19,12 +19,7 @@ module DataMigrate
       def migrate
         DataMigrate::DataMigrator.assure_data_schema_table
         target_version = ENV["VERSION"] ? ENV["VERSION"].to_i : nil
-        if (Rails::VERSION::MAJOR == 5 && Rails::VERSION::MINOR == 2) ||
-          Rails::VERSION::MAJOR == 6
-          DataMigrate::MigrationContext.new(migrations_paths).migrate(target_version)
-        else
-          DataMigrate::DataMigrator.migrate(migrations_paths, ENV["VERSION"] ? ENV["VERSION"].to_i : nil)
-        end
+        DataMigrate::MigrationContext.new(migrations_paths).migrate(target_version)
       end
 
       def abort_if_pending_migrations(migrations, message)
