+++ spec/spec_helper.rb
@@ -20,24 +20,8 @@ RSpec.configure do |config|
     if example.metadata[:no_override]
     else
       @prev_data_migrations_path = DataMigrate.config.data_migrations_path
-      if Rails::VERSION::MAJOR == 6
-        DataMigrate.configure do |config|
-          config.data_migrations_path = "spec/db/6.0"
-        end
-      elsif Rails::VERSION::MAJOR == 5
-        if Rails::VERSION::MINOR == 2
-          DataMigrate.configure do |config|
-            config.data_migrations_path = "spec/db/data"
-          end
-        else
-          DataMigrate.configure do |config|
-            config.data_migrations_path = "spec/db/5.0"
-          end
-        end
-      else
-        DataMigrate.configure do |config|
-          config.data_migrations_path = "spec/db/5.0"
-        end
+      DataMigrate.configure do |config|
+        config.data_migrations_path = "spec/db/6.0"
       end
     end
   end
